requirements.txt
flask
numpy
pillow
rasterio
fiona
shapely
ultralytics